beginning python rewrite
the upcoming pants roadmap includes:
 * i'm going to add an abstract interpreter into the compilation pipeline for
   static analysis and optimization.
 * i'm going to stop targeting c with label pointers, which means i need to
   start outputting real functions and expect tail recursion optimization in
   the c compiler, or i should target llvm.
 * i want to fix a bunch of stuff in the parser by writing it by hand (which
   is this commit)

a lot of this stuff is serious rewrite work, and i'm sick of the visitor
design pattern. so, i'm just going to start over in python for "rapid"
prototyping.

i'll retranslate back to c with what i've learned once i'm done.

this commit provides the beginnings of a by-hand parser that implements a
bunch of stuff i punted on last time, such as
 * comments
 * multiline strings
 * string escapes
 * variable base integers and floats
 * line/column referencing
 * argument ordering
 * etc

also, it should be pretty fast due to memoization